Create PointChartView abstraction; Create ScatterChartView subclass
